chatbot-backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# Google Gemini API (RAG backend logic)
google-generativeai==0.8.3

# ChatKit SDK (event streaming and frontend integration)
# Note: ChatKit Python SDK to be added when available
# chatkit-sdk==0.1.0  # Placeholder

# Vector database (Qdrant)
qdrant-client==1.12.0

# Database (Neon Postgres)
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# API contract validation
openapi-spec-validator==0.7.1

# Environment variables
python-dotenv==1.0.1

# HTTP client
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
